requirements.txt
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dash==1.19.0
dash-bootstrap-components==0.11.1
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.4
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.21.0
openpyxl==3.0.6
packaging==20.4
pandas~=1.2.5
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.14.3
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pyzmq==19.0.2
retrying==1.3.3
scipy==1.7.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
tabula-py==2.2.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==2.0.1

scikit-learn~=0.24.1
nltk~=3.5
pdfplumber~=0.5.25
umap~=0.1.1
umap_learn~=0.3.10

pdftotext~=2.1.5